refactor to helper methods
